(let's try again)
Right now test includes ALL the words from the deck. Make it so I can choose how many words I want the test to include. After this add a special function/button to take the test with all the words in the deck included. If all answers are correct - mark the deck as "learned".